feat(git-utils): add support for additional commit message keywords
Added support for additional commit message keywords in the git-utils.js file. The patchWords array now includes the keywords "style", "docs", "test", and "chore" in addition to the existing keywords for fixing and closing issues. This change enhances the classification of commit messages in the versioning logic.

Refactor the patchWords assignment to use case-insensitive regular expressions for better keyword matching.